+++ src/Guilded.Commands/CommandEvent.cs
@@ -1,9 +1,5 @@
-using System;
 using System.Collections.Generic;
-using System.Threading.Tasks;
-using Guilded.Base;
 using Guilded.Base.Content;
-using Guilded.Base.Embeds;
 using Guilded.Base.Events;
 
 namespace Guilded.Commands;
@@ -15,7 +11,7 @@ namespace Guilded.Commands;
 /// <seealso cref="MessageEvent" />
 /// <seealso cref="FailedCommandEvent" />
 /// <seealso cref="CommandAttribute" />
-public class CommandEvent
+public class CommandEvent : MessageEvent
 {
     #region Properties
 
@@ -33,13 +29,7 @@ public class CommandEvent
     /// <para>Even if we invoke <q>config items add</q> command, the root command will always remain <q>config</q>.</para>
     /// </example>
     /// <value>Command context</value>
-    public RootCommandContext RootCommand { get; }
-
-    /// <summary>
-    /// Gets the message event that invoked the command.
-    /// </summary>
-    /// <value>Message creation event</value>
-    public MessageEvent MessageEvent => RootCommand.MessageEvent;
+    public RootCommandEvent RootCommand { get; }
 
     /// <summary>
     /// Gets the prefix that has been used on the command.
@@ -72,41 +62,6 @@ public class CommandEvent
     public IEnumerable<string> Arguments { get; }
     #endregion
 
-    #region MessageEvent properties
-    /// <inheritdoc cref="MessageEvent{T}.Message" />
-    public Message Message => MessageEvent.Message;
-
-    /// <inheritdoc cref="ChannelContent{TId, TServer}.ServerId" />
-    public HashId? ServerId => Message.ServerId;
-
-    /// <inheritdoc cref="ChannelContent{TId, TServer}.ChannelId" />
-    public Guid ChannelId => Message.ChannelId;
-
-    /// <inheritdoc cref="Message.Content" />
-    public string? Content => Message.Content;
-
-    /// <inheritdoc cref="ChannelContent{TId, TServer}.CreatedBy" />
-    public HashId CreatedBy => Message.CreatedBy;
-
-    /// <inheritdoc cref="Message.CreatedByWebhook" />
-    public Guid? CreatedByWebhook => Message.CreatedByWebhook;
-
-    /// <inheritdoc cref="ChannelContent{TId, TServer}.CreatedAt" />
-    public DateTime CreatedAt => Message.CreatedAt;
-
-    /// <inheritdoc cref="Message.IsReply" />
-    public bool IsReply => Message.IsReply;
-
-    /// <inheritdoc cref="Message.ReplyMessageIds" />
-    public IList<Guid>? ReplyMessageIds => Message.ReplyMessageIds;
-
-    /// <inheritdoc cref="Message.Embeds" />
-    public IList<Embed>? Embeds => Message.Embeds;
-
-    /// <inheritdoc cref="Message.IsPrivate" />
-    public bool IsPrivate => Message.IsPrivate;
-    #endregion
-
     #endregion
 
     #region Constructors
@@ -116,77 +71,7 @@ public class CommandEvent
     /// <param name="context">The the root-level command that was used</param>
     /// <param name="commandName">The name of the command that was used</param>
     /// <param name="arguments">The array of string arguments that were given to the command</param>
-    public CommandEvent(RootCommandContext context, string commandName, IEnumerable<string> arguments) =>
+    public CommandEvent(RootCommandEvent context, string commandName, IEnumerable<string> arguments) : base(context.MessageEvent.Message, context.MessageEvent.ServerId) =>
         (RootCommand, CommandName, Arguments) = (context, commandName, arguments);
-    #endregion
-
-    #region Additional
-    /// <inheritdoc cref="Message.CreateMessageAsync(MessageContent)" />
-    public async Task<Message> CreateMessageAsync(MessageContent message) =>
-        await Message.CreateMessageAsync(message).ConfigureAwait(false);
-
-    /// <inheritdoc cref="Message.CreateMessageAsync(string)" />
-    public async Task<Message> CreateMessageAsync(string content) =>
-        await Message.CreateMessageAsync(content).ConfigureAwait(false);
-
-    /// <inheritdoc cref="Message.CreateMessageAsync(string, bool, bool)" />
-    public async Task<Message> CreateMessageAsync(string content, bool isPrivate = false, bool isSilent = false) =>
-        await Message.CreateMessageAsync(content, isPrivate, isSilent).ConfigureAwait(false);
-
-    /// <inheritdoc cref="Message.CreateMessageAsync(string, Guid[])" />
-    public async Task<Message> CreateMessageAsync(string content, params Guid[] replyMessageIds) =>
-        await Message.CreateMessageAsync(content, replyMessageIds).ConfigureAwait(false);
-
-    /// <inheritdoc cref="Message.CreateMessageAsync(string, bool, bool, Guid[])" />
-    public async Task<Message> CreateMessageAsync(string content, bool isPrivate = false, bool isSilent = false, params Guid[] replyTo) =>
-        await Message.CreateMessageAsync(content, isPrivate, isSilent, replyTo).ConfigureAwait(false);
-
-    /// <inheritdoc cref="Message.CreateMessageAsync(Embed[])" />
-    public async Task<Message> CreateMessageAsync(params Embed[] embeds) =>
-        await Message.CreateMessageAsync(embeds).ConfigureAwait(false);
-
-    /// <inheritdoc cref="Message.CreateMessageAsync(bool, bool, Guid[], Embed[])" />
-    public async Task<Message> CreateMessageAsync(bool isPrivate = false, bool isSilent = false, Guid[]? replyTo = null, params Embed[] embeds) =>
-        await Message.CreateMessageAsync(isPrivate, isSilent, replyTo, embeds).ConfigureAwait(false);
-
-    /// <inheritdoc cref="Message.CreateMessageAsync(string, Embed[])" />
-    public async Task<Message> CreateMessageAsync(string content, params Embed[] embeds) =>
-        await Message.CreateMessageAsync(content, embeds).ConfigureAwait(false);
-
-    /// <inheritdoc cref="Message.CreateMessageAsync(string, bool, bool, Guid[], Embed[])" />
-    public async Task<Message> CreateMessageAsync(string content, bool isPrivate = false, bool isSilent = false, Guid[]? replyTo = null, params Embed[] embeds) =>
-        await Message.CreateMessageAsync(content, isPrivate, isSilent, replyTo, embeds).ConfigureAwait(false);
-
-    /// <inheritdoc cref="Message.ReplyAsync(string)" />
-    public async Task<Message> ReplyAsync(string content) =>
-        await Message.ReplyAsync(content).ConfigureAwait(false);
-
-    /// <inheritdoc cref="Message.ReplyAsync(string, bool, bool)" />
-    public async Task<Message> ReplyAsync(string content, bool isPrivate = false, bool isSilent = false) =>
-        await Message.ReplyAsync(content, isPrivate, isSilent).ConfigureAwait(false);
-
-    /// <inheritdoc cref="Message.ReplyAsync(bool, bool, Embed[])" />
-    public async Task<Message> ReplyAsync(bool isPrivate = false, bool isSilent = false, params Embed[] embeds) =>
-        await Message.ReplyAsync(isPrivate, isSilent, embeds).ConfigureAwait(false);
-
-    /// <inheritdoc cref="Message.ReplyAsync(string, bool, bool, Embed[])" />
-    public async Task<Message> ReplyAsync(string content, bool isPrivate = false, bool isSilent = false, params Embed[] embeds) =>
-        await Message.ReplyAsync(content, isPrivate, isSilent, embeds).ConfigureAwait(false);
-
-    /// <inheritdoc cref="Message.UpdateAsync(string)" />
-    public async Task<Message> UpdateAsync(string content) =>
-        await Message.UpdateAsync(content).ConfigureAwait(false);
-
-    /// <inheritdoc cref="Message.DeleteAsync" />
-    public async Task DeleteAsync() =>
-        await Message.DeleteAsync().ConfigureAwait(false);
-
-    /// <inheritdoc cref="Message.AddReactionAsync(uint)" />
-    public async Task<Reaction> AddReactionAsync(uint emoteId) =>
-        await Message.AddReactionAsync(emoteId).ConfigureAwait(false);
-
-    // /// <inheritdoc cref="Message.RemoveReactionAsync(uint)" />
-    // public async Task RemoveReactionAsync(uint emoteId) =>
-    //     await Message.RemoveReactionAsync(emoteId).ConfigureAwait(false);
     #endregion
 }